message is +dump +blink +UART +USB
Removed other debug loop or interpreter code
(still in comments at the moment).

Shortened the wait time after cold boot.

To catch the earliest part, UART is easier
since it can remain connected when cycling
power, except for .UF2 upload, which is
intolerant of this on the reference machine
(CP2104).

Presumably, the TX and RX pins are passing
enough current to mess with the USB bootloader.

	modified:   camelforth-a.c

On branch dvlp-aa-