Results from latest UF2 already committed
watchdog not very suitable for automated use,
to recover from a problem with 'getchar()'
spitting out an endless stream of 0xFF, after
behaving correctly for more than half an hour
after power on and regular camelforth interpreter
action and activity.

Will remove the automatic use of the watchdog,
as this is now seen as of not much value.

The one time it'd be helpful is if the terminal
threatens to flood with endless output, once the
so-called 'race condition' of getchar() begins.

More useful would be to be able to probe the
Forth stack after this kind of an ABEND, before
what appears to be a mandatory restart by cycling
power to the RP2040 Pico target board.

	modified:   results_debugging.txt

On branch dvlp-bb-Error-debug-