no_flash commit ee - FUNCTIONAL NOW
No problems expected for this demo code,
other than what happens to forth when
the flwrite word is used.

Seems somewhat function on brief test, as
flwrite was used twice in a row without using
COLD or power cycling.

Just a first test - writes to flash do succeed.

0x10040000 is where to look. ;)

ok 10040000 10 - 3 DUMP

	modified:   forth.c

On branch copy-to-ram-a