STABLE flwrite erase and new keyboard-await blinker.
copy_to_ram mode tested good.

Very brief testing.  Blinks every 10 seconds (or so)
while waiting for keyboard input.

More like 'stable' with a very small 's' there.

Expecting it to 'suddenly not work as it had been'
due to luck alone. ;)

The RPi Pico RP2040 board is curiously presenting to Linux
as /dev/ttyACM1 (and the transition was noted). Not sure
what's happening there.

There is no /dev/ttyACM0 present (now) but that was the
device until yesterday.

	modified:   .gitignore
	modified:   17-camelforth-a.d/pico-examples/camelforth-a/forth/forth.c
	modified:   doc/results_debugging.txt

On branch erase_sector-a