STABLE reading word 1023 bytes read into SRAM
The resulting SRAM buffer image (at 0x2000DF44 currently)
is null-terminated ('\0'). 1023 bytes of information
plus the null is the 1024 bytes of this buffer.

Intent is to eventually use this SRAM buffer,
populated by keyboard input, as a holding area
for forth source code.

Possibly: an 'flaccept' type word that parses
forth out of QSPI flashROM space 0x10000000
plus some significant offset.

Intermediate goal is to simply populate QSPI
flash with this captured source code, for later
(possible) parsing by a putative flaccept word
(and friends).

	modified:   forth.c

On branch erase_sector-a-rdng-flacc-a